% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwlabel3d.R
\name{bwlabel3d}
\alias{bwlabel3d}
\title{Binary segmentation in 3d}
\usage{
bwlabel3d(img)
}
\arguments{
\item{img}{A 3d array. x is considered as a binary image, whose pixels of value 0 are considered as background ones and other pixels as foreground ones.}
}
\value{
A Grayscale 3d array, containing the labelled version of x.
}
\description{
Binary segmentation in 3d
}
\author{
Fabian Scheipl, Volker Schmid
}
