% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmoments3d.R
\name{cmoments3d}
\alias{cmoments3d}
\title{Computes moments from image objects}
\usage{
cmoments3d(mask, ref)
}
\arguments{
\item{mask}{a labeled stack as returned from bwlabel3d}

\item{ref}{the original image stack}
}
\value{
a matrix with the moments of the objects in the stack
}
\description{
Computes intensity-weighted centers of objects and their mass (sum of intensities) and size.
}
\author{
Volker Schmid
}
