% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colorsinclasses.R
\name{colors.in.classes}
\alias{colors.in.classes}
\title{Title Compute colors in classes distribution}
\usage{
colors.in.classes(classes, color1, color2 = NULL, mask = array(TRUE,
  dim(classes)), N = max(classes, na.rm = TRUE), thresh1 = NULL,
  thresh2 = NULL, sd1 = 2, sd2 = 2, col1 = "green", col2 = "red",
  test = FALSE, plot = TRUE)
}
\arguments{
\item{classes}{Image of classes}

\item{color1}{Image of first color}

\item{color2}{Image of second color}

\item{mask}{Image mask}

\item{N}{Maximum number of classes}

\item{thresh1}{Threshold for first color image}

\item{thresh2}{Threshold for second color image}

\item{sd1}{For automatic threshold, that is: mean(color1)+sd1*sd(color1)}

\item{sd2}{For automatic threshold of color2}

\item{col1}{Name of color 1}

\item{col2}{Name of color 2}

\item{test}{Compute tests}

\item{plot}{Plot barplots}
}
\value{
Table of classes with color 1 (and 2)
}
\description{
Title Compute colors in classes distribution
}

