% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterImage3d.R
\name{filterImage3d}
\alias{filterImage3d}
\title{Apply filter to 3D images}
\usage{
filterImage3d(img, filter = "var", window, z.scale = 1, silent = FALSE)
}
\arguments{
\item{img}{is a 3d arrary representing an image.}

\item{filter}{is the filter to be applied. Options: var: Variance filter.}

\item{window}{half size of window; i.e. window=1 uses a window of 3 voxels in each direction.}

\item{z.scale}{ratio of voxel dimension in x/y direction and z direction.}

\item{silent}{Logical. If FALSE, information on progress will be printed.}
}
\value{
Multi-dimensional array of filtered image data.
}
\description{
A filter is applied to a 3D array representing an image. So far only variance filters are supported.
}
