% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_choose.R
\name{folder.choose}
\alias{folder.choose}
\title{Choose a folder interactively}
\usage{
folder.choose()
}
\value{
A character vector of length one giving the folder path.
}
\description{
Choose a folder interactively by choosing a file in that folder.
}

