% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/image.R
\name{image}
\alias{image}
\title{image for microscopy}
\usage{
image(x, col = "grey", ...)
}
\arguments{
\item{x}{Image, 2D Matrix}

\item{col}{Color, "grey", "red" ("r"), "green" ("g") or "blue" ("b")}

\item{...}{other parameters for graphics::image}
}
\value{
no return
}
\description{
image for microscopy
}

