% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Kcross3D.R
\name{K.cross.3D}
\alias{K.cross.3D}
\title{K-function cross-type in 3D}
\usage{
K.cross.3D(X, Y, Z, X2, Y2, Z2, psz = 25, width = 1, intensity = NULL,
  intensity2 = NULL, parallel = FALSE, bw = 0.01)
}
\arguments{
\item{X}{X coordinate of first observed point pattern in microns.}

\item{Y}{Y coordinate}

\item{Z}{Z coordinate}

\item{X2}{X coordinate of second observed point pattern}

\item{Y2}{Y coordinate}

\item{Z2}{Z coordinate}

\item{psz}{pointsize used for discetization. Smaller values are more precise, but need more computation time.}

\item{width}{maximum distance}

\item{intensity}{intensity of first pattern. Only if \deqn{\lambda(s)!=\lambda}.}

\item{intensity2}{intensity of second pattern}

\item{parallel}{Logical. Can we use parallel computing?}

\item{bw}{bandwith}
}
\value{
a list of breaks and counts.
}
\description{
Calculates an estimate of the cross-type K-function for a multitype point pattern.
}

