% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3cross3d.R
\name{L.cross.3D}
\alias{L.cross.3D}
\title{L-function cross-type in 3d}
\usage{
L.cross.3D(X, Y, Z, X2, Y2, Z2, psz = 25, width = 1, intensity = NULL,
  intensity2 = NULL, parallel = FALSE, verbose = FALSE)
}
\arguments{
\item{X}{X coordinate of first observed point pattern in microns.}

\item{Y}{Y coordinate}

\item{Z}{Z coordinate}

\item{X2}{X coordinate of second observed point pattern}

\item{Y2}{Y coordinate}

\item{Z2}{Z coordinate}

\item{psz}{pointsize used for discetization. Smaller values are more precise, but need more computation time.}

\item{width}{maximum distance}

\item{intensity}{intensity of first pattern. Only if \deqn{\lambda(s)!=\lambda}.}

\item{intensity2}{intensity of second pattern}

\item{parallel}{Logical. Can we use parallel computing?}

\item{verbose}{Plot verbose information}
}
\value{
a list of breaks and counts.
}
\description{
Calculates an estimate of the cross-type L-function for a multitype point pattern.
}

