% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mexican_hat_brush.R
\name{mexican.hat.brush}
\alias{mexican.hat.brush}
\title{Mexican hat brush
to use with filter2}
\usage{
mexican.hat.brush(n = 7, sigma2 = 1)
}
\arguments{
\item{n}{size of brush}

\item{sigma2}{standard deviation}
}
\value{
brush
}
\description{
Mexican hat brush
to use with filter2
}
