% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestclassdistances_c.R
\name{nearestClassDistance3}
\alias{nearestClassDistance3}
\title{Title Find distance to next neighbour of a specific class}
\usage{
nearestClassDistance3(coord, img, class, voxelsize, step = NULL)
}
\arguments{
\item{coord}{coordinate of relevant voxel}

\item{img}{image array of classes}

\item{class}{class to find}

\item{voxelsize}{vector of length three. size of voxel in X-/Y-/Z-direction}
}
\value{
distance to nearest voxel of class "class"
}
\description{
Title Find distance to next neighbour of a specific class
}

