% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_neighbour_distance.R
\name{nearestClassDistances}
\alias{nearestClassDistances}
\title{Title Find all distances to next neighbour of all classes}
\usage{
nearestClassDistances(img, voxelsize, classes = 7, cores = 1)
}
\arguments{
\item{img}{image array of classes}

\item{voxelsize}{vector of length three. size of voxel in X-/Y-/Z-direction}

\item{classes}{Number of classes}

\item{cores}{Number of cores available for parallel computing}
}
\value{
array with distances
}
\description{
Title Find all distances to next neighbour of all classes
}

