% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nearest_class_distances.R
\name{plotNearestClassDistances}
\alias{plotNearestClassDistances}
\title{Title Plot nearest class distances}
\usage{
plotNearestClassDistances(longlist, method, classes = dim(longlist)[1],
  ylim = c(0, 1), qu = 0.01)
}
\arguments{
\item{longlist}{array with distances as produces by nearest.class.distances()}

\item{method}{"boxplot", "min" or "quantile"}

\item{classes}{number of classes, default=7}

\item{ylim}{limits for distances, default=c(0,1)}

\item{qu}{quantile for method="quantile", default=0.01}
}
\value{
plots
}
\description{
Title Plot nearest class distances
}

