% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readClassTIF.R
\name{readClassTIF}
\alias{readClassTIF}
\title{Read TIF file with classes}
\usage{
readClassTIF(file, n = 7)
}
\arguments{
\item{file}{file}

\item{n}{number of classes}
}
\value{
array
}
\description{
Read TIF file with classes
}

