% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testColoc.R
\name{testColoc}
\alias{testColoc}
\title{Permutation Test for cross-type nearest neighbor distances}
\usage{
testColoc(im1, im2, hres = 0.102381, vres = 0.25, B = 999,
  alternative = "less", returnSample = TRUE, ...)
}
\arguments{
\item{im1}{image stack as returned by preprocessing}

\item{im2}{image stack as returned by preprocessing}

\item{hres}{horizontal resolution of the stacks}

\item{vres}{vertical resolution of the stacks}

\item{B}{number of permutations to generate}

\item{alternative}{alternative hypothesis ("less" to test H0:Colocalization )}

\item{returnSample}{return sampled null distibution}

\item{...}{additional arguments for papply}
}
\value{
a list with the p.value, the observed weighted mean of the cNN-distances
}
\description{
Permutation Test for cross-type nearest neighbor distances
}
\author{
Fabian Scheipl
}
