% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{img}
\alias{img}
\title{image for microscopy}
\usage{
img(x, col = "grey", low = 0, up = NULL, ...)
}
\arguments{
\item{x}{Image, 2D or 3D Matrix}

\item{col}{Color, either a character ("grey", "greyinvert", "red" ("r"), "green" ("g") or "blue" ("b"), rgb" for 3D matrices) or a function}

\item{low}{minimal value of shown intensity}

\item{up}{maximal value of shown intensity}

\item{...}{other parameters for graphics::image}
}
\value{
no return
}
\description{
image for microscopy
}

