% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity3d.R
\name{intensity3D}
\alias{intensity3D}
\title{Intensity of a 3d Dataset or a Model}
\usage{
intensity3D(X, Y, Z, bw = NULL, psz = 25, kernel = "Square")
}
\arguments{
\item{X}{X coordinate}

\item{Y}{Y coordinate}

\item{Z}{Z coordinate}

\item{bw}{bandwidth}

\item{psz}{pointsize used for discretization (large: fast, but not precise)}

\item{kernel}{"Square" or "Uniform"}
}
\value{
3d Array
}
\description{
Computing the intensity of a 3d point pattern using kernel smoothing.
}
