% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestclassdistances_c.R
\name{nearestClassDistances}
\alias{nearestClassDistances}
\title{Find all distances to next neighbor of all classes}
\usage{
nearestClassDistances(
  img,
  voxelsize = NULL,
  size = NULL,
  classes = 7,
  maxdist = NULL,
  silent = FALSE,
  cores = 1
)
}
\arguments{
\item{img}{Image array of classes}

\item{voxelsize}{Real size of voxels in microns.}

\item{size}{Real size of image in microns. Either size or voxelsize must be given.}

\item{classes}{Number of classes}

\item{maxdist}{Maximum distance to consider}

\item{silent}{Remain silent?}

\item{cores}{Number of cores available for parallel computing}
}
\value{
array with distances
}
\description{
Find all distances to next neighbor of all classes
}
