% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestclassdistances.R
\name{nearestClassDistances}
\alias{nearestClassDistances}
\title{Find all distances to next neighbour of all classes}
\usage{
nearestClassDistances(img, voxelsize, classes = 7, cores = 1)
}
\arguments{
\item{img}{Image array of classes}

\item{voxelsize}{Vector of length three. Size of voxel in X-/Y-/Z-direction}

\item{classes}{Number of classes}

\item{cores}{Number of cores available for parallel computing}
}
\value{
array with distances
}
\description{
Find all distances to next neighbour of all classes
}

