% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readbmp.R
\name{readBMP}
\alias{readBMP}
\title{Read bitmap files}
\usage{
readBMP(file)
}
\arguments{
\item{file}{A character vector of file names or URLs.}
}
\value{
Returns a matrix with BMP data as integer.
}
\description{
Read 2D grey-value BMP files
}
\examples{
bi<-readBMP(system.file("extdata/V.bmp",package="bioimagetools"))
image(bi,col=grey(seq(1,0,length=100)))
}
\author{
Volker J. Schmid
}
