% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readtif.R
\name{readTIF}
\alias{readTIF}
\title{Read tif stacks}
\usage{
readTIF(file = file.choose(), native = FALSE, as.is = FALSE,
  channels = NULL)
}
\arguments{
\item{file}{Name of the file to read from. Can also be an URL.}

\item{native}{determines the image representation - if FALSE (the default) then the result is an array, if TRUE then the result is a native raster representation (suitable for plotting).}

\item{as.is}{attempt to return original values without re-scaling where possible}

\item{channels}{number of channels}
}
\value{
3d or 4d array
}
\description{
Read tif stacks
}
\examples{
kringel <- readTIF(system.file("extdata","kringel.tif",package="bioimagetools"))
img(kringel)
}
